pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcsmt"
version = "0.3.0.1"
authors = [
  { name="ECSDevs", email="psbash@groupmail.com" },
]
description = "Simple file synchronization service (clients in the same package)"
readme = "README.md"
requires-python = ">=3.0"
license = {file = "LICENCE"}
maintainers = [
  {name = "originalFactor", email = "2438926613@qq.com" }
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Operating System :: OS Independent",
]
dependencies = [
  "dnspython",
  "httpx",
  "watchdog"
]

[project.urls]
"Homepage" = "https://github.com/ECSDevs/MCSMT"
"Bug Tracker" = "https://github.com/ECSDevs/MCSMT/issues"
"Support Me" = "https://afdian.net/a/mcsmt_transfer_tool"